splitting argparsing functions away from log manipulation functions
