+++ cancat2candump
@@ -1,4 +1,4 @@
 #!/usr/bin/env python2
 
-from cancat.utils.log import cancat2candump
-cancat2candump()
+from cancat.utils.cancat2candump import main
+main()
