+++ candump2cancat
@@ -1,4 +1,4 @@
 #!/usr/bin/env python2
 
-from cancat.utils.log import candump2cancat
-candump2cancat()
+from cancat.utils.candump2cancat import main
+main()
